+++ examples/react-apollo/src/sign-in/email-password.tsx
@@ -1,7 +1,7 @@
 import { useState } from 'react'
 import { useNavigate } from 'react-router-dom'
 
-import { Button, Modal, TextInput } from '@mantine/core'
+import { Button, Divider, Modal, TextInput } from '@mantine/core'
 import { showNotification } from '@mantine/notifications'
 import { useSignInEmailPassword } from '@nhost/react'
 
@@ -78,7 +78,6 @@ export const EmailPassword: React.FC = () => {
           value={email}
           onChange={(e) => setEmail(e.target.value)}
           placeholder="Email Address"
-          size="lg"
           autoFocus
           style={{ marginBottom: '0.5em' }}
         />
@@ -87,15 +86,17 @@ export const EmailPassword: React.FC = () => {
           onChange={(e) => setPassword(e.target.value)}
           placeholder="Password"
           type="password"
-          size="lg"
           style={{ marginBottom: '0.5em' }}
         />
-
         <Button fullWidth onClick={signIn}>
           Sign in
         </Button>
         <AuthLink link="/sign-in/forgot-password" variant="white">
           Forgot password?
+        </AuthLink>
+        <Divider />
+        <AuthLink link="/sign-in" variant="white">
+          &#8592; Other Sign-in Options
         </AuthLink>
       </>
     )