+++ packages/stripe-graphql-js/dev/server.ts
@@ -1,0 +1,18 @@
+import { Context, createStripeGraphQLServer } from '../src/index'
+
+const isAllowed = (stripeCustomerId: string, context: Context) => {
+  const { isAdmin } = context
+
+  if (isAdmin) {
+    return true
+  }
+
+  return false
+}
+
+const server = createStripeGraphQLServer({
+  isAllowed,
+  graphiql: true
+})
+
+server.start()