+++ packages/stripe-graphql-js/src/schema/payment-method-card-three-d-secure-usage.ts
@@ -1,0 +1,9 @@
+import { builder } from '../builder'
+
+builder.objectType('StripePaymentMethodCardThreeDSecureUsage', {
+  fields: (t) => ({
+    supported: t.exposeBoolean('supported', {
+      description: `Whether 3D Secure is supported on this card.`
+    })
+  })
+})