+++ packages/core/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @nhost/core
 
+## 0.8.0
+
+### Minor Changes
+
+- 8e4d790b: Sign up with an email and a security key.
+
+## 0.7.7
+
+### Patch Changes
+
+- 9eb78e06: Add `workos` social provider
+
 ## 0.7.6
 
 ### Patch Changes