+++ packages/hasura-storage-js/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @nhost/hasura-storage-js
 
+## 0.6.2
+
+### Patch Changes
+
+- Updated dependencies [8e4d790b]
+  - @nhost/core@0.8.0
+
+## 0.6.1
+
+### Patch Changes
+
+- Updated dependencies [9eb78e06]
+  - @nhost/core@0.7.7
+
 ## 0.6.0
 
 ### Minor Changes